// src/adapter/generated-types.ts
export interface Board {
  version: number;
  bump: number;
  round_id: number;
  round_duration: number;
  start_slot: BN;
  end_slot: BN;
  strike_pending_usd_amount: BN;
  strike_usd_amount: BN;
  strike_btc_amount: BN;
  strike_last_trigger_round_id: number;
  strike_reserve_usd_amount: BN;
  strike_reserve_btc_amount: BN;
  rotor_arm_counter: BN;
  strike_token_amount: BN;
  strike_reserve_token_amount: BN;
  reserved: number[];
}
// ...
export interface SatrushConfig {
  version: number;
  bump: number;
  owner_authority: PublicKey;
  admin_authority: PublicKey;
  game_authority: PublicKey;
  fee_recipient: PublicKey;
  usd_mint: PublicKey;
  btc_mint: PublicKey;
  strike_fee_bps: number;
  epoch_fee_bps: number;
  one_btc_fee_bps: number;
  sats_vault_round_fee_bps: number;
  vault_exit_fee_bps: number;
  protocol_fee_bps: number;
  unclaimed_hashrate_bps: number;
  min_deploy_usd_amount: BN;
  epoch_vault_iteration_duration: BN;
  deployment_settle_grace_duration: BN;
  strike_trigger_modulus: number;
  buybacks_fee_bps: number;
  token_mint: PublicKey;
  reserved: number[];
}

export interface SatsClaimed {
  authority: PublicKey;
  claimed_shares: BN;
  btc_received: BN;
  claimed_hashrate: BN;
  claimed_token_shares: BN;
  tokens_received: BN;
}

export interface SatsVault {
  version: number;
  bump: number;
  btc_amount: BN;
  btc_shares: BN;
  leftovers: BN;
  reserved: number[];
}

export interface TileStake {
  stake: BN;
  deploy_count: number;
}

export interface TokenClaimed {
  authority: PublicKey;
  claimed_shares: BN;
  tokens_received: BN;
  claimed_sats_shares: BN;
  btc_received: BN;
  claimed_hashrate: BN;
}

export interface TokenVault {
  version: number;
  bump: number;
  token_amount: BN;
  token_shares: BN;
  leftovers: BN;
  reserved: number[];
}